move code to backend; use decorator
