+++ backends/alephalpha_api.py
@@ -5,6 +5,7 @@ import aleph_alpha_client
 import anthropic
 import backends
 from backends import ModelSpec, Model
+from backends.utils import ensure_messages_format
 
 logger = backends.get_logger(__name__)
 
@@ -28,6 +29,7 @@ class AlephAlphaModel(backends.Model):
         self.client = client
 
     @retry(tries=3, delay=0, logger=logger)
+    @ensure_messages_format
     def generate_response(self, messages: List[Dict]) -> Tuple[Any, Any, str]:
         """
         :param messages: for example
