+++ backends/anthropic_api.py
@@ -4,6 +4,8 @@ import anthropic
 import backends
 import json
 
+from backends.utils import ensure_messages_format
+
 logger = backends.get_logger(__name__)
 
 NAME = "anthropic"
@@ -24,6 +26,7 @@ class AnthropicModel(backends.Model):
         self.client = client
 
     @retry(tries=3, delay=0, logger=logger)
+    @ensure_messages_format
     def generate_response(self, messages: List[Dict]) -> Tuple[str, Any, str]:
         """
         :param messages: for example
