+++ backends/cohere_api.py
@@ -2,6 +2,7 @@ from typing import List, Dict, Tuple, Any
 from retry import retry
 import cohere
 import backends
+from backends.utils import ensure_messages_format
 import json
 
 logger = backends.get_logger(__name__)
@@ -26,6 +27,7 @@ class CohereModel(backends.Model):
         self.client = client
 
     @retry(tries=3, delay=0, logger=logger)
+    @ensure_messages_format
     def generate_response(self, messages: List[Dict]) -> Tuple[str, Any, str]:
         """
         :param messages: for example
