+++ backends/openai_api.py
@@ -4,6 +4,7 @@ from retry import retry
 import json
 import openai
 import backends
+from backends.utils import ensure_messages_format
 
 logger = backends.get_logger(__name__)
 
@@ -36,6 +37,7 @@ class OpenAIModel(backends.Model):
         self.client = client
 
     @retry(tries=3, delay=0, logger=logger)
+    @ensure_messages_format
     def generate_response(self, messages: List[Dict]) -> Tuple[str, Any, str]:
         """
         :param messages: for example
