+++ backends/openai_compatible_api.py
@@ -6,6 +6,8 @@ import openai
 import backends
 import httpx
 
+from backends.utils import ensure_messages_format
+
 logger = backends.get_logger(__name__)
 
 NAME = "generic_openai_compatible"
@@ -41,6 +43,7 @@ class GenericOpenAIModel(backends.Model):
         self.client = client
 
     @retry(tries=3, delay=0, logger=logger)
+    @ensure_messages_format
     def generate_response(self, messages: List[Dict]) -> Tuple[str, Any, str]:
         """
         :param messages: for example
