+++ clemgame/clemgame.py
@@ -20,45 +20,6 @@ stdout_logger = clemgame.get_logger("benchmark.run")
 GAMES_TO_IGNORE = ["hellogame", "chatgame"]
 
 
-def ensure_alternating_roles(messages: List[Dict], cull_system_message: bool = True) -> List[Dict]:
-    """
-    The messages format assumes alternating roles of user and assistant. This method checks, if this constraint
-    is satisfied. If this is not the case and there are consecutive user or assistant messages,
-    then these are merged into a single one.
-
-    :param messages: to be checked
-    :return: a new messages object with the alternating roles ensured
-    """
-    _messages = copy.deepcopy(messages)
-
-    if cull_system_message:
-        if _messages[0]['role'] == "system" and not _messages[0]["content"]:
-            del _messages[0]
-
-    def is_same_role(msg1, msg2):
-        return msg1["role"] == msg2["role"]
-
-    delimiter = "\n\n"
-
-    def join_content(msg1, msg2):
-        return f"{msg1['content']}{delimiter}{msg2['content']}"
-
-    # combine consecutive user messages:
-    for msg_idx, message in enumerate(_messages):
-        if msg_idx == 0:
-            continue
-        prev_message = _messages[msg_idx - 1]
-        if is_same_role(prev_message, message):
-            warn_msg = (f"Found consecutive role assignments. These will be merged into one:\n"
-                        f"{prev_message}\n"
-                        f"{message}")
-            logger.warn(warn_msg)
-            prev_message['content'] = join_content(prev_message, message)
-            del _messages[msg_idx]
-
-    return _messages
-
-
 class Player(abc.ABC):
     """
     A participant of a game. A player can respond via a custom implementation, human input or a language model:
@@ -78,7 +39,6 @@ class Player(abc.ABC):
 
     def __call__(self, messages: List[Dict], turn_idx) -> Tuple[Any, Any, str]:
         call_start = datetime.now()
-        messages = ensure_alternating_roles(messages)
         prompt = messages
         response = dict()
         if isinstance(self.model, CustomResponseModel):
